backend/requirements.txt
aioserial==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.2
attrs==25.3.0
bcrypt==4.0.1
Beaker==1.13.0
beautifulsoup4==4.14.2
blinker==1.9.0
blivet==3.12.1
blivet-gui==2.6.0
Brlapi==0.8.6
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorlog==6.9.0
crcmod==1.7
crypt_r==3.13.1
cryptography==45.0.7
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
distro==1.9.0
dnf==4.23.0
dnspython==2.8.0
docopt==0.6.2
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
fastapi-slim==0.116.1
fedora-third-party==0.10
file-magic==0.4.1
filelock==3.20.0
fonttools==4.59.2
fros==1.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
greenlet==3.2.4
groq==0.31.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hidapi==0.14.0.post4
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanize==4.13.0
i3ipc==2.2.1
idna==3.10
iso639==0.1.4
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==0.3.76
langsmith==0.4.27
langtable==0.0.69
libcomps==0.1.23.post1
libdnf==0.74.0
liquidctl==1.15.0
louis==3.33.0
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
mercurial==7.1.1
meson==1.9.0
motor==3.7.1
mpmath==1.3.0
msgpack==1.1.1
narwhals==2.5.0
netifaces==0.11.0
networkx==3.5
nftables==0.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
nwg-panel==0.10.5
oauthlib==3.3.1
olefile==0.47
openai==2.6.1
opencv-python==4.12.0.88
orjson==3.11.3
packaging==25.0
pandas==2.3.2
passlib==1.7.4
Paste==3.10.1
peewee==3.18.2
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
ply==3.11
productmd==1.48
progressbar2==4.5.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pwquality==1.4.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.13
pycairo==1.25.1
pycparser==2.23
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pyenchant==3.3.0
PyGObject==3.50.0
pyinotify==0.9.6
PyJWT==2.10.1
pykickstart==3.67
pymongo==4.15.3
pynvim==0.6.0
pyOpenSSL==25.2.0
pyparsing==3.2.4
pyparted==3.13.0
pyserial==3.5
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-linux-procfs==0.7.3
python-meh==0.52
python-multipart==0.0.20
python-pam==2.0.2
python-utils==3.9.1
python-xlib==0.33
pytz==2025.2
pyudev==0.24.3
pyusb==1.3.1
pywal==3.3.0
pywal16==3.8.9
pywalfox==2.7.4
pyxdg==0.28
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.16.9-build/kernel-6.16.9/linux-6.16.9-200.fc42.x86_64/tools/net/ynl
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
requests-ftp==0.3.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.27.1
rpm==4.20.1
rpmautospec==0.8.3
rpmautospec-core==0.1.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setproctitle==1.3.7
setuptools==80.9.0
simpleline==1.9.0
six==1.17.0
smbus==1.1.post2
smmap==5.0.2
sniffio==1.3.1
sos==4.10.0
soupsieve==2.8
starlette==0.47.3
sympy==1.14.0
systemd-python==235
Tempita==0.6.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
triton==3.5.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
udiskie==2.5.8
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
xkbregistry==0.3
xmltodict==1.0.0
zstandard==0.25.0
